cheri: Manually apply fixes for FixPtrToULongCast warnings
Manually add fixes for the FixPtrToULongCast warnings where
clang-tidy isn't able to fix them automatically due to macro
expansions.

Explicitly ignore warnings in fs/btrfs/ because this file system
plays tricks with pointers that need a different fix.

Ignore some warnings that need a different fix.

+++ arch/riscv/include/asm/elf.h
@@ -91,7 +91,7 @@ do {								\
 	 * COMPAT ELFCLASS32.					\
 	 */							\
 	NEW_AUX_ENT(AT_SYSINFO_EHDR,				\
-		(elf_addr_t)(ulong)current->mm->context.vdso);	\
+		(elf_addr_t)(uintptr_t)current->mm->context.vdso);	\
 	NEW_AUX_ENT(AT_L1I_CACHESIZE,				\
 		get_cache_size(1, CACHE_TYPE_INST));		\
 	NEW_AUX_ENT(AT_L1I_CACHEGEOMETRY,			\
