+++ sound/soc/codecs/lpass-rx-macro.c
@@ -600,7 +600,7 @@ enum {
 	.info = rx_macro_iir_filter_info, \
 	.get = rx_macro_get_iir_band_audio_mixer, \
 	.put = rx_macro_put_iir_band_audio_mixer, \
-	.private_value = (unsigned long)&(struct wcd_iir_filter_ctl) { \
+	.private_value = (uintptr_t)&(struct wcd_iir_filter_ctl) { \
 		.iir_idx = iidx, \
 		.band_idx = bidx, \
 		.bytes_ext = {.max = RX_MACRO_IIR_FILTER_SIZE, }, \
